requirements.txt
pandas>=2.1.0
numpy>=1.26.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
great-expectations>=0.18.0
python-dotenv>=1.0.0